make repartition as internal method of DataFrame and Series
