+++ docs/flow/modin/distributed/dataframe/pandas.rst
@@ -12,10 +12,6 @@ from_partitions
 ---------------
 .. autofunction:: modin.distributed.dataframe.pandas.from_partitions
 
-repartition
------------
-.. autofunction:: modin.distributed.dataframe.pandas.repartition
-
 Example
 -------
